pyproject.toml
[tool.poetry]
name = "site-bit-loader"
version = "0.1.0"
description = ""
authors = ["Weit145 <weit.dima@yandex.ru>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.2"
alembic = "^1.16.5"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pydantic = {extras = ["email"], version = "^2.11.9"}
sqlalchemy = "^2.0.43"
aiosqlite = "^0.21.0"
pydantic-settings = "^2.10.1"
pyjwt = "^2.10.1"
passlib = {extras = ["argon2", "bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.1.1"
python-multipart = "^0.0.20"
argon2-cffi = "^25.1.0"
bcrypt = "^4.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
black = "^25.9.0"
ruff = "^0.14.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
    "migrations/",
    ".git",
    "__pycache__",
    ".ruff_cache",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "F841",  # unused variable
]
isort = {known-first-party = ["app"]}

[tool.ruff.format]
indent-style = "space"
quote-style = "double"

[tool.ruff.lint.pydocstyle]
convention = "google"